Move ArmPreviewDragAndDrop.js to urdfScripts
